requirements.txt
anthropic>=0.39.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
reportlab>=4.0
python-docx>=1.0
lxml>=4.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
